feat(realtime): add console channel
