+++ app/controllers/shared/api.php
@@ -166,7 +166,7 @@ App::init(function ($utopia, $request, $project) {
                 throw new Exception('JWT authentication is disabled for this project', 501);
             }
             break;
-        
+
         default:
             throw new Exception('Unsupported authentication route');
             break;
@@ -187,7 +187,7 @@ App::shutdown(function ($utopia, $request, $response, $project, $events, $audits
     /** @var bool $mode */
 
     if (!empty($events->getParam('event'))) {
-        if(empty($events->getParam('eventData'))) {
+        if (empty($events->getParam('eventData'))) {
             $events->setParam('eventData', $response->getPayload());
         }
 
@@ -206,10 +206,10 @@ App::shutdown(function ($utopia, $request, $response, $project, $events, $audits
 
         if ($project->getId() !== 'console') {
             $payload = new Document($response->getPayload());
-            $target = Realtime::fromPayload($events->getParam('event'), $payload);
+            $target = Realtime::fromPayload($events->getParam('event'), $payload, $project);
 
             Realtime::send(
-                $project->getId(),
+                $target['projectId'] ?? $project->getId(),
                 $response->getPayload(),
                 $events->getParam('event'),
                 $target['channels'],
